Cranelift: remove support for core in favor of restricted-std
The core (i.e. no-std support) feature was unmaintained and not
compiling. This commit replaces that feature with a new restricted-std
feature that can be used in environments supporting only a minimal
subset of the standard library.

The restricted-std feature introduces very few changes to the codebase
other than removing #![no_std] and enabling the use of std instead of
core and alloc. In particular code requiring advances std features (such
as codegen timing which uses thread-local storage) were already
feature-gated by #[cfg(not(feature = "std"))].